+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added inherited methods to class docs.
 - Added data structure mixins to the docs.
-- Added `data`, `to_data`, and `from_data` to `compas.geometry.Polyhedron`
+- Added `data` and `from_data` to `compas.geometry.Polyhedron`
 
 ### Changed